refactor: genesis processing with tests
