+++ src/genesis/db/table_manager.py
@@ -1,77 +1,92 @@
-from typing import Any, Generator, Tuple
+import itertools
+from contextlib import contextmanager
+from enum import Enum
+from typing import Any, Generator, List, Tuple
 
 from psycopg import Connection
-from psycopg.errors import UniqueViolation
-
-from src.genesis.db.types import DBTypes
 
 
-def table_exists(db_conn: Connection, table: str) -> bool:
-    with db_conn.cursor() as db:
-        res_db_execute = db.execute(
-            f"""
-                SELECT EXISTS (
-                    SELECT FROM pg_tables WHERE
-                        schemaname = 'app' AND
-                        tablename  = '{table}'
-                )
-            """
-        ).fetchone()
-
-        assert res_db_execute is not None
-
-        return res_db_execute[0]
+class DBTypes(Enum):
+    text = "text"
+    numeric = "numeric"
+    interface = "public.app_enum_0f6c2478ba"
 
 
 class TableManager:
-    _db_conn: Connection
-    _table: str
-    _columns: Tuple[Tuple[str, DBTypes], ...]  # [(<name>, <type>)]
-    _indexes: Tuple[str, ...]
-    _schema: str = "app"
+    def __init__(
+        self,
+        db_conn: Connection,
+        table: str,
+        columns: Tuple[Tuple[str, DBTypes], ...],
+        indexes: Tuple[str, ...],
+        schema: str = "app",
+    ):
+        self.db_conn = db_conn
+        self.table = table
+        self.columns = columns
+        self.indexes = indexes
+        self.schema = schema
 
-    def __init__(self, db_conn: Connection):
-        self._db_conn = db_conn
+    def get_column_names(self) -> Generator[str, Any, None]:
+        return (name for name, _ in self.columns)
 
-    @classmethod
-    def get_column_names(cls) -> Generator[str, Any, None]:
-        return (name for name, _ in cls._columns)
-
-    @classmethod
-    def select_query(cls) -> str:
-        return f"""
-            SELECT {",".join(cls.get_column_names())} FROM {cls._table}
+    def select_query(self, column_names: List[str]) -> str:
+        res = self.db_conn.execute(
+            f"""
+            SELECT {",".join(column_names)} FROM {self.table}
         """
+        ).fetchall()
+        return list(itertools.chain(*res))
 
-    def _ensure_table(self):
-        with self._db_conn.cursor() as db:
+    def ensure_table(self):
+        with self.db_conn.cursor() as db:
             db.execute(
                 f"""
-                CREATE TABLE IF NOT EXISTS {self._schema}.{self._table} (
-                    {", ".join([f"{name} {type_.value}" for name, type_ in self._columns])}
+                CREATE TABLE IF NOT EXISTS {self.schema}.{self.table} (
+                    {", ".join([f"{name} {type_.value}" for name, type_ in self.columns])}
                 );
                 -- TODO: psycopg break out of transaction
-                -- CREATE INDEX CONCURRENTLY ON {self._schema}.{self._table} ({",".join(self._indexes)})
+                -- CREATE INDEX CONCURRENTLY ON {self.schema}.{self.table} ({",".join(self.indexes)})
             """
             )
-            self._db_conn.commit()
+            self.db_conn.commit()
             # TODO error checking / handling (?)
 
-    def _drop_table(self, cascade: bool = False):
+    def drop_table(self, cascade: bool = False):
         cascade_clause = ""
         if cascade:
             cascade_clause = "CASCADE"
 
-        with self._db_conn.cursor() as db:
+        with self.db_conn.cursor() as db:
             db.execute(
                 f"""
-                DROP TABLE IF EXISTS {self._schema}.{self._table} {cascade_clause};
+                DROP TABLE IF EXISTS {self.schema}.{self.table} {cascade_clause};
             """
             )
-            self._db_conn.commit()
+            self.db_conn.commit()
             # TODO error checking / handling (?)
 
-    @classmethod
-    def _extract_id_from_unique_violation_exception(cls, e: UniqueViolation) -> str:
-        # Extract which ID was violated from UniqueViolation exception
-        return str(e).split("(")[2].split(")")[0]
+    def table_exists(self, table: str) -> bool:
+        with self.db_conn.cursor() as db:
+            res_db_execute = db.execute(
+                f"""
+                    SELECT EXISTS (
+                        SELECT FROM pg_tables WHERE
+                            schemaname = 'app' AND
+                            tablename  = '{table}'
+                    )
+                """
+            ).fetchone()
+
+            assert res_db_execute is not None
+
+            return res_db_execute[0]
+
+    @contextmanager
+    def db_copy(self):
+        with self.db_conn.cursor() as db:
+            with db.copy(
+                f'COPY {self.table} ({",".join(self.get_column_names())}) FROM STDIN'
+            ) as copy:
+                yield copy
+        self.db_conn.commit()
