--- src/genesis/db/types.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-
-class DBTypes(Enum):
-    text = "text"
-    numeric = "numeric"
