--- src/genesis/helpers/__init__.py
@@ -1,1 +1,0 @@
-from .field_enums import *  # noqa: F401
