+++ src/genesis/processing/balances.py
@@ -1,9 +1,9 @@
 from typing import List
 
-from db.table_manager import DBTypes, TableManager
+from src.genesis.db.table_manager import DBTypes, TableManager
 from psycopg import Connection
 
-from utils.loggers import get_logger
+from src.genesis.utils.loggers import get_logger
 
 _logger = get_logger(__name__)
