+++ tests/e2e/db/test_table_manager.py
@@ -2,8 +2,7 @@ import sys
 import unittest
 from pathlib import Path
 
-from src.genesis.db.table_manager import TableManager, table_exists
-from src.genesis.db.types import DBTypes
+from src.genesis.db.table_manager import TableManager, DBTypes
 from tests.helpers.clients import TestWithDBConn
 
 src_path = Path(__file__).parent.parent.parent.parent.absolute()
@@ -35,22 +34,22 @@ class TestTableManager(TestWithDBConn):
             )
 
     def test__ensure_table(self) -> None:
-        exists = table_exists(self.db_conn, self.test_table)
+        exists = self.table_manager.table_exists(self.test_table)
         self.assertFalse(exists)
 
-        self.table_manager._ensure_table()
-        exists = table_exists(self.db_conn, self.test_table)
+        self.table_manager.ensure_table()
+        exists = self.table_manager.table_exists(self.test_table)
         self.assertTrue(exists)
 
     # TODO: test cascade
     # NB: _drop_table test depends on _ensure_table's correctness
     def test__drop_table(self) -> None:
-        self.table_manager._ensure_table()
-        exists = table_exists(self.db_conn, self.test_table)
+        self.table_manager.ensure_table()
+        exists = self.table_manager.table_exists(self.db_conn, self.test_table)
         self.assertTrue(exists)
 
         self.table_manager._drop_table()
-        exists = table_exists(self.db_conn, self.test_table)
+        exists = self.table_manager.table_exists(self.db_conn, self.test_table)
         self.assertFalse(exists)
