+++ tests/e2e/observers/test_balances.py
@@ -2,47 +2,12 @@ import copy
 import unittest
 from threading import Lock
 from typing import List
-from unittest.mock import patch
 
-import pytest
-from reactivex.scheduler import ThreadPoolScheduler
-
-from src.genesis.genesis import Genesis
 from src.genesis.helpers.field_enums import GenesisBalances
-from src.genesis.observers import (
-    GenesisBalancesManager,
-    GenesisBalancesObserver,
-    native_balances_keys_path,
-)
-from src.genesis.state import Balance, Coin
+from src.genesis.processing.balances import BalanceManager
+
 from tests.helpers.clients import TestWithDBConn
 from tests.helpers.genesis_data import test_bank_state_balances, test_genesis_data
-
-
-class TestNativeBalanceObserver(TestWithDBConn):
-    def test_subscribe_to(self) -> None:
-        completed = False
-        test_genesis = Genesis(**test_genesis_data)
-        actual_entries: List[Balance] = []
-        expected_entries = [
-            (native_balances_keys_path, Balance(**b))
-            for b in test_genesis_data["app_state"]["bank"]["balances"]
-        ]
-
-        def on_next(balance: Balance):
-            actual_entries.append(balance)
-
-        def on_completed():
-            nonlocal completed
-            completed = True
-
-        test_balance_observer = GenesisBalancesObserver(
-            on_next=on_next, on_completed=on_completed
-        )
-        test_balance_observer.subscribe_to(test_genesis.source)
-
-        self.assertTrue(completed)
-        self.assertListEqual(expected_entries, actual_entries)
 
 
 class TestBalanceManager(TestWithDBConn):
@@ -65,41 +30,28 @@ class TestBalanceManager(TestWithDBConn):
         # Clean DB to prevent interaction with other tests
         self.reinit_db()
 
-        expected_balances: List[Balance] = Balance.from_dict_list(
-            test_bank_state_balances
-        )
-        scheduler = ThreadPoolScheduler(2)
-        lock = Lock()
-        lock.acquire()
+        test_manager = BalanceManager(self.db_conn)
+        test_manager.process_genesis(test_genesis_data)
 
-        def on_completed():
-            actual_balances: List[Balance] = self.collect_actual_balances()
-            self.check_balances(expected_balances, actual_balances)
-            lock.release()
-
-        test_manager = GenesisBalancesManager(self.db_conn, on_completed=on_completed)
-        test_manager.observe(Genesis(**test_genesis_data).source, scheduler=scheduler)
-
-        # Lock returns false if times-out
-        assert lock.acquire(True, 5)
+        actual_balances: [dict] = self.collect_actual_balances()
+        self.check_balances(test_bank_state_balances, actual_balances)
 
     def collect_actual_balances(self):
         actual_balances = []
 
         with self.db_conn.cursor() as db:
             for address in [b["address"] for b in test_bank_state_balances]:
-                balance = Balance(**{"address": address})
+                balance = {"address": address}
+                balance["coins"] = []
 
                 for row in db.execute(
-                    GenesisBalances.select_where(f"account_id = '{address}'")
+                        GenesisBalances.select_where(f"account_id = '{address}'")
                 ).fetchall():
-                    balance.coins.append(
-                        Coin(
-                            **{
-                                "amount": int(row[GenesisBalances.amount.value]),
-                                "denom": row[GenesisBalances.denom.value],
-                            }
-                        )
+                    balance["coins"].append(
+                        {
+                            "amount": int(row[GenesisBalances.amount.value]),
+                            "denom": row[GenesisBalances.denom.value],
+                        }
                     )
 
                 actual_balances.append(balance)
@@ -111,81 +63,25 @@ class TestBalanceManager(TestWithDBConn):
             found_balance = False
             actual_balance = None
             for actual_balance_ in actual_balances:
-                if actual_balance_.address == expected_balance.address:
+                if actual_balance_["address"] == expected_balance["address"]:
                     found_balance = True
                     actual_balance = actual_balance_
                     break
 
             self.assertTrue(found_balance)
-            self.assertEqual(expected_balance.address, actual_balance.address)
+            self.assertEqual(expected_balance["address"], actual_balance["address"])
 
-            for expected_coin in expected_balance.coins:
+            for expected_coin in expected_balance["coins"]:
                 found_coin = False
                 actual_coin = None
-                for actual_coin_ in actual_balance.coins:
-                    if actual_coin_.denom == expected_coin.denom:
+                for actual_coin_ in actual_balance["coins"]:
+                    if actual_coin_["denom"] == expected_coin["denom"]:
                         found_coin = True
                         actual_coin = actual_coin_
                         break
 
                 self.assertTrue(found_coin)
                 self.assertEqual(expected_coin, actual_coin)
-
-        # TODO: check for extra stuff in actual_balances (?)
-
-    @patch("logging.Logger.warning")
-    def test_observe_with_duplicate_values(self, logger_warning_mock):
-        # Clean DB to prevent interaction with other tests
-        self.reinit_db()
-
-        duplicate_message = "Duplicate balance occurred"
-
-        # Insert first set of balances to DB
-        test_manager = GenesisBalancesManager(self.db_conn)
-        test_manager.observe(Genesis(**test_genesis_data).source)
-
-        # Try to insert same set again
-        second_test_manager = GenesisBalancesManager(self.db_conn)
-        second_test_manager.observe(Genesis(**test_genesis_data).source)
-
-        n_min_calls = 4
-        assert logger_warning_mock.call_count == n_min_calls
-        for mock_call in logger_warning_mock.mock_calls:
-            assert duplicate_message in mock_call.args[0]
-
-    def test_observe_with_duplicate_values_with_errors(self):
-        # Clean DB to prevent interaction with other tests
-        self.reinit_db()
-
-        current_bank_state_balances = [
-            {
-                "address": "addr123",
-                "coins": [
-                    {"amount": 123, "denom": "a-token"},
-                    {"amount": 457, "denom": "b-token"},
-                ],
-            },
-        ]
-
-        # Create variation of test data without overwriting the original dict
-        current_test_genesis_data = copy.deepcopy(test_genesis_data)
-        current_test_genesis_data["app_state"]["bank"][
-            "balances"
-        ] = current_bank_state_balances
-
-        # Insert first set of balances to DB
-        test_manager = GenesisBalancesManager(self.db_conn)
-        test_manager.observe(Genesis(**test_genesis_data).source)
-
-        # Insert duplicate entry with different balance
-        test_manager = GenesisBalancesManager(self.db_conn)
-
-        with pytest.raises(RuntimeError) as e:
-            test_manager.observe(Genesis(**current_test_genesis_data).source)
-        assert (
-            "Balance for addr123-b-token in DB (456) is different from genesis (457)"
-            in str(e)
-        )
 
 
 if __name__ == "__main__":
